+++ modules/cloudfront/custom/headers.tf
@@ -28,6 +28,7 @@ resource "null_resource" "update_cloudfront_headers" {
   }
 
   provisioner "local-exec" {
+    interpreter = ["bash", "-ec"]
     command = <<EOF
 export AWS_DEFAULT_REGION=${data.aws_region.current.name}
 export AWS_REGION=${data.aws_region.current.name}