+++ modules/console-ui/console_ui.tf
@@ -94,6 +94,7 @@ resource "null_resource" "trigger_console_ui_upgrade" {
   }
 
   provisioner "local-exec" {
+    interpreter = ["bash", "-ec"]
     command = <<EOF
 export AWS_DEFAULT_REGION=${data.aws_region.current.name}
 export AWS_REGION=${data.aws_region.current.name}