+++ modules/mosaic-titiler/providers.tf
@@ -2,6 +2,9 @@ terraform {
   required_providers {
     aws = {
       source = "hashicorp/aws"
+      configuration_aliases = [
+        aws.east
+      ]
     }
   }
 }